[+ @seddy ] namespace the github stuff
We want to insulate most of this app from having to integrate with
github hence lets namespace stuff now to keep it clean
