+++ lib/shipping_agent/deployer.rb
@@ -1,0 +1,9 @@
+module ShippingAgent
+  module Deployer
+    extend self
+
+    def notify(url)
+      puts url
+    end
+  end
+end
